+++ src/migrations/2015_03_06_135351_create_inventory_supplier_tables.php
@@ -11,7 +11,7 @@ class CreateInventorySupplierTables extends Migration
     public function up()
     {
         Schema::create('suppliers', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->timestamps();
 
             $table->string('name');
@@ -29,11 +29,11 @@ class CreateInventorySupplierTables extends Migration
         });
 
         Schema::create('inventory_suppliers', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->timestamps();
 
-            $table->integer('inventory_id')->unsigned();
-            $table->integer('supplier_id')->unsigned();
+            $table->bigInteger('inventory_id')->unsigned();
+            $table->bigInteger('supplier_id')->unsigned();
 
             $table->foreign('inventory_id')->references('id')->on('inventories')
                 ->onUpdate('restrict')
